fix(jest): refactored test dir
