--- frontend/.jest/__mocks__/data.mock.js
@@ -1,6 +1,0 @@
-export const user = {
-    name: 'Karin',
-    userSettings: {
-        feedbackLifespan: 2
-    }
-}
