--- frontend/.jest/__mocks__/lottie-player.mock.js
@@ -1,3 +1,0 @@
-module.exports = {
-  LottiePlayer: () => <div data-testid="lottie-player-mock"></div>,
-};
