+++ frontend/jest/mocks/context.mock.js
@@ -1,0 +1,16 @@
+import * as AppContext from '@contexts/app.context';
+import {
+    user,
+} from './data.mock.js';
+
+export const contextValues = {
+    isLoggedIn: true,
+    setIsLoggedIn: jest.fn(),
+    user: user,
+    setUser: jest.fn(),
+    isCookieConsentOpen: true,
+    setIsCookieConsentOpen: jest.fn(),
+};
+
+jest.spyOn(AppContext, 'useAppContext')
+	.mockImplementation(() => contextValues);
