+++ frontend/jest/mocks/data.mock.js
@@ -1,0 +1,6 @@
+export const user = {
+    name: 'Karin',
+    userSettings: {
+        feedbackLifespan: 2
+    }
+}
