+++ frontend/jest/mocks/lottie-player.mock.js
@@ -1,0 +1,3 @@
+module.exports = {
+  LottiePlayer: () => <div data-testid="lottie-player-mock"></div>,
+};
