+++ frontend/jest/setup.js
@@ -1,0 +1,61 @@
+import '@testing-library/jest-dom';
+
+global.XMLHttpRequest = undefined;
+
+global.resizeWindow = (width, height) => {
+  window.innerWidth = width;
+  window.innerHeight = height;
+  window.dispatchEvent(new Event('resize'));
+};
+
+// Mock IntersectionObserver
+class IntersectionObserver {
+  observe = jest.fn();
+  disconnect = jest.fn();
+  unobserve = jest.fn();
+}
+
+Object.defineProperty(window, 'IntersectionObserver', {
+  writable: true,
+  configurable: true,
+  value: IntersectionObserver,
+});
+
+Object.defineProperty(global, 'IntersectionObserver', {
+  writable: true,
+  configurable: true,
+  value: IntersectionObserver,
+});
+
+jest.mock('next/image', () => ({
+  __esModule: true,
+  default: () => {
+    return 'Next image stub';
+  },
+}));
+
+jest.mock('next/router', () => ({
+  push: jest.fn().mockImplementation(() => Promise.resolve()),
+  back: jest.fn(),
+  events: {
+    on: jest.fn(),
+    off: jest.fn(),
+  },
+  query: {},
+  route: '',
+  beforePopState: jest.fn(() => null),
+  __esModule: true,
+  useRouter: () => ({
+    route: '/',
+    pathname: '',
+    query: {},
+    asPath: '',
+    push: jest.fn().mockImplementation(() => Promise.resolve()),
+    events: {
+      on: jest.fn(),
+      off: jest.fn(),
+    },
+    beforePopState: jest.fn(() => null),
+    prefetch: jest.fn(() => null),
+  }),
+}));
