--- frontend/src/services/__tests__/api-service.test.jsx
@@ -1,18 +1,0 @@
-import { handleError } from '@services/api-service';
-
-describe('Api service', () => {
-  it('Test handleError to throw the error', () => {
-    const error = {
-      response: {
-        status: 401,
-        data: {
-          message: 'Unauthorized',
-        },
-      },
-      config: {
-        url: 'url',
-      },
-    };
-    expect(() => handleError(error).toThrow(error));
-  });
-});
