+++ frontend/tsconfig.base.json
@@ -32,7 +32,7 @@
       "@utils/*": ["src/utils/*"],
       "@middlewares/*": ["src/utils/middlewares/*"],
       "@public/*": ["public/*"],
-      "@jestRoot/*": [".jest/*"],
+      "@jestRoot/*": ["jest/*"],
       "@data-contracts/*": ["src/data-contracts/*"]
     }
   }
